+++ src/main/java/com/educatech/service/impl/LessonService.java
@@ -14,6 +14,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
@@ -31,9 +32,24 @@ public class LessonService implements ILessonService {
      * @return DTO de respuesta de la lección guardada.
      */
     @Override
+    @Transactional
     public LessonResponseDTO saveLesson(LessonRequestDTO lesson) {
+        if (lesson.getCourseId() == null || lesson.getCourseId() <= 0) {
+            throw new IllegalArgumentException("Course ID must be provided and greater than zero.");
+        }
+        Course course = courseRepository.findById(lesson.getCourseId())
+                .orElseThrow(() -> new CourseNotFoundException("Course not found with id: " + lesson.getCourseId()));
+
+        lessonRepository.getLessonByTitleAndCourse(lesson.getTitle(), course).ifPresent(
+                l -> {
+                    throw new IllegalArgumentException("A lesson with the same title already exists in this course.");
+                }
+        );
+
         Lesson lessonToSave = lessonMapper.toEntity(lesson);
+        lessonToSave.setContent(lesson.getContent());
         Lesson savedLesson = lessonRepository.save(lessonToSave);
+
         return lessonMapper.toResponseDTO(savedLesson);
     }
 
@@ -44,6 +60,7 @@ public class LessonService implements ILessonService {
      * @return Página de DTOs de respuesta de lecciones.
      */
     @Override
+    @Transactional(readOnly = true)
     public Page<LessonResponseDTO> getAllLessons(Pageable pageable) {
         return lessonRepository.findAll(pageable).map(lessonMapper::toResponseDTO);
     }
@@ -55,7 +72,11 @@ public class LessonService implements ILessonService {
      * @return DTO de respuesta de la lección.
      */
     @Override
+    @Transactional(readOnly = true)
     public LessonResponseDTO getLessonById(Long idLesson) {
+        if (idLesson == null || idLesson <= 0) {
+            throw new IllegalArgumentException("Lesson ID must be provided and greater than zero.");
+        }
         return lessonMapper.toResponseDTO(this.getLessonEntityById(idLesson));
     }
 
@@ -67,15 +88,39 @@ public class LessonService implements ILessonService {
      * @return DTO de respuesta de la lección actualizada.
      */
     @Override
+    @Transactional
     public LessonResponseDTO updateLesson(Long idLesson, LessonRequestDTO lessonWithUpdates) {
+        if (idLesson == null || idLesson <= 0) {
+            throw new IllegalArgumentException("Lesson ID must be provided and greater than zero.");
+        }
+        if (lessonWithUpdates.getCourseId() == null || lessonWithUpdates.getCourseId() <= 0) {
+            throw new IllegalArgumentException("Course ID must be provided and greater than zero.");
+        }
+
+        Course course = courseRepository.findById(lessonWithUpdates.getCourseId())
+                .orElseThrow(() -> new CourseNotFoundException("Course not found with id: " + lessonWithUpdates.getCourseId()));
+
+        lessonRepository.getLessonByTitleAndCourse(lessonWithUpdates.getTitle(), course)
+                .ifPresent(lesson -> {
+                    if (!lesson.getId().equals(idLesson)) {
+                        throw new IllegalArgumentException("A lesson with the same title already exists in this course.");
+                    }
+                });
+
+        // verifica si la lección existe
         Lesson existingLesson = this.getLessonEntityById(idLesson);
+
+        // EL problema de este metodo es que si quiero actualizar el NOMBRE y el CONTENIDO de la lección
+        // y NO el curso, me da error porque la lección ya está asociada a ese curso.
+        // Lo que se podría hacer es permitir que no se actualice el curso si no se quiere cambiar.
+        // lo mejor es darlo para un PUT completo, manejarlo con un PATCH para actualizaciones parciales
+
+//        if (existingLesson.getCourse().getId().equals(course.getId())) {
+//            throw new IllegalArgumentException("The lesson is already associated with the specified course.");
+//        }
 
         existingLesson.setTitle(lessonWithUpdates.getTitle());
         existingLesson.setContent(lessonWithUpdates.getContent());
-
-        Course course = courseRepository.findById(lessonWithUpdates.getCourseId())
-                .orElseThrow(() -> new LessonNotFoundException("Course not found with id: " + lessonWithUpdates.getCourseId()));
-
         existingLesson.setCourse(course);
 
         Lesson updatedLesson = lessonRepository.save(existingLesson);
@@ -88,7 +133,12 @@ public class LessonService implements ILessonService {
      * @param idLesson ID de la lección a eliminar.
      */
     @Override
+    @Transactional
     public void deleteLesson(Long idLesson) {
+        if (idLesson == null || idLesson <= 0) {
+            throw new IllegalArgumentException("Lesson ID must be provided and greater than zero.");
+        }
+        // Verifica si la lección existe
         Lesson lessonToDelete = this.getLessonEntityById(idLesson);
         lessonRepository.delete(lessonToDelete);
     }
@@ -100,7 +150,11 @@ public class LessonService implements ILessonService {
      * @return Lista de DTOs de respuesta de lecciones.
      */
     @Override
+    @Transactional(readOnly = true)
     public List<LessonResponseDTO> getLessonsByCourse(Long idCourse) {
+        if (idCourse == null || idCourse <= 0) {
+            throw new IllegalArgumentException("Course ID must be provided and greater than zero.");
+        }
         Course course = courseRepository.findById(idCourse)
                 .orElseThrow(() -> new CourseNotFoundException("Course not found with id: " + idCourse));
 
@@ -110,13 +164,7 @@ public class LessonService implements ILessonService {
                 .toList();
     }
 
-    /**
-     * Método auxiliar para obtener una entidad de lección por su ID.
-     * Lanza una excepción si no se encuentra.
-     *
-     * @param idLesson ID de la lección.
-     * @return Entidad de lección.
-     */
+    // Método privado para obtener una lección por su ID o lanzar una excepción si no existe
     private Lesson getLessonEntityById(Long idLesson) {
         return lessonRepository.findById(idLesson)
                 .orElseThrow(() -> new LessonNotFoundException("Lesson not found with id: " + idLesson));
